Cast temperature to int

+++ Fagelmatare-AVR/fagelmatare.ino
@@ -306,7 +306,7 @@ void loop() {
        * According to sensor DC Electrical Characteristics: see datasheet table 1
        * For thermometer in range -10 to 85 °C ±0.5
        */
-      Serial.print(ceil(temperature/0.5)*5);
+      Serial.print(int(ceil(temperature/0.5)*5));
       Serial.print('\0');
       return;
     }
